// src/app/shared/types/api-types/bire-unmount-order-dto.interface.ts
export interface BireUnmountOrderDTO {
  familyCode?: string;
  variantCode?: string;
  structureType?: string;
  unmountCode?: string;
  unmFamilyCode?: string;
  unmVariantCode?: string;
  unmStructureType?: string;
  unmUnmountCode?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}